添加了确认删除

+++ web/admin/goods_list.jsp
@@ -83,10 +83,25 @@
 
 					</p>
 					<a class="btn btn-success" href="${pageContext.request.contextPath}/admin/goods_editshow?id=${g.id }& pageNumber=${p.pageNumber}&type=${type}">修改</a>
-					<a class="btn btn-danger" href="${pageContext.request.contextPath}/admin/goods_delete?id=${g.id }&pageNumber=${p.pageNumber}&type=${type}">删除</a>
+					<a class="btn btn-danger" href="${pageContext.request.contextPath}/admin/goods_delete?id=${g.id }&pageNumber=${p.pageNumber}&type=${type}" onclick="confirmDelete(${g.id}, ${p.pageNumber}, '${type}')">删除</a>
 				</td>
 			</tr>
 		</c:forEach>
+		<script>
+			function confirmDelete(id, pageNumber, type) {
+				// 弹出提示框，并获取用户点击的按钮
+				var confirmed = confirm("确认删除吗？");
+
+				// 如果用户点击了确定按钮，执行删除操作
+				if (confirmed) {
+					// 构造删除商品的 URL
+					var deleteUrl = "${pageContext.request.contextPath}/admin/goods_delete?id=" + id + "&pageNumber=" + pageNumber + "&type=" + type;
+
+					// 执行删除操作
+					window.location.href = deleteUrl;
+				}
+			}
+		</script>
 
 
 	</table>
